improved test of device fakeLaserWithMotor several cleanups to laser-related devices
