+++ .ci/initial-cache.gh.linux.cmake
@@ -63,7 +63,7 @@ set(ENABLE_yarpmod_SDLJoypad ON CACHE BOOL "")
 set(ENABLE_yarpmod_batteryClient ON CACHE BOOL "")
 set(ENABLE_yarpmod_batteryWrapper ON CACHE BOOL "")
 set(ENABLE_yarpmod_upowerBattery ON CACHE BOOL "")
-set(ENABLE_yarpmod_Rangefinder2DWrapper ON CACHE BOOL "")
+set(ENABLE_yarpmod_rangefinder2D_nws_yarp ON CACHE BOOL "")
 set(ENABLE_yarpmod_multipleanalogsensorsserver ON CACHE BOOL "")
 set(ENABLE_yarpmod_multipleanalogsensorsclient ON CACHE BOOL "")
 set(ENABLE_yarpmod_multipleanalogsensorsremapper ON CACHE BOOL "")
