+++ src/devices/fakeLaserWithMotor/tests/fakeLaserWithMotorTest.cpp
@@ -4,10 +4,12 @@
  */
 
 #include <yarp/dev/IRangefinder2D.h>
+#include <yarp/dev/IPositionControl.h>
 #include <yarp/os/Network.h>
 #include <yarp/dev/PolyDriver.h>
 #include <yarp/dev/WrapperSingle.h>
 #include <yarp/dev/tests/IRangefinder2DTest.h>
+#include <yarp/dev/tests/IPositionControlTest.h>
 
 #include <catch.hpp>
 #include <harness.h>
@@ -16,9 +18,9 @@ using namespace yarp::dev;
 using namespace yarp::sig;
 using namespace yarp::os;
 
-TEST_CASE("dev::FakeLaserTest", "[yarp::dev]")
+TEST_CASE("dev::FakeLaserWithMotorTest", "[yarp::dev]")
 {
-    YARP_REQUIRE_PLUGIN("fakeLaser", "device");
+    YARP_REQUIRE_PLUGIN("fakeLaserWithMotor", "device");
 
     Network::setLocalMode(true);
 
@@ -26,19 +28,22 @@ TEST_CASE("dev::FakeLaserTest", "[yarp::dev]")
     {
         PolyDriver fakelaserdev;
         IRangefinder2D* irng = nullptr;
+        IPositionControl* ipos = nullptr;
 
         ////////"Checking opening polydriver"
         {
             Property las_cfg;
-            las_cfg.put("device", "fakeLaser");
+            las_cfg.put("device", "fakeLaserWithMotor");
             las_cfg.put("test", "use_constant");
             las_cfg.put("const_distance", 0.5);
             REQUIRE(fakelaserdev.open(las_cfg));
             REQUIRE(fakelaserdev.view(irng));
+            REQUIRE(fakelaserdev.view(ipos));
         }
 
         //execute tests
         yarp::dev::tests::exec_iRangefinder2D_test_1(irng);
+        yarp::dev::tests::exec_iPositionControl_test_1(ipos);
 
         //"Close all polydrivers and check"
         CHECK(fakelaserdev.close());
