+++ src/devices/laserFromExternalPort/laserFromExternalPort.cpp
@@ -34,16 +34,16 @@ using namespace yarp::dev;
 YARP_LOG_COMPONENT(LASER_FROM_EXTERNAL_PORT, "yarp.devices.laserFromExternalPort")
 
 /*
-yarpdev --device Rangefinder2DWrapper --subdevice laserFromExternalPort \
+yarpdev --device rangefinder2D_nws_yarp --subdevice laserFromExternalPort \
 --SENSOR::input_ports_name "(/port1 /port2)" \
 --TRANSFORM_CLIENT::local /LaserFromExternalPort/tfClient \
 --TRANSFORM_CLIENT::remote /transformServer \
 --TRANSFORMS::src_frames "(/frame1 /frame2)" \
 --TRANSFORMS::dst_frame /output_frame
---period 10 \
+--period 0.01 \
 --name /outlaser:o
 
-yarpdev --device Rangefinder2DWrapper --subdevice laserFromExternalPort \
+yarpdev --device rangefinder2D_nws_yarp --subdevice laserFromExternalPort \
 --SENSOR::min_angle 0
 --SENSOR::max_angle 360
 --SENSOR::resolution 0.5
@@ -52,7 +52,7 @@ yarpdev --device Rangefinder2DWrapper --subdevice laserFromExternalPort \
 --TRANSFORM_CLIENT::remote /transformServer \
 --TRANSFORMS::src_frames "(/frame1 /frame2)" \
 --TRANSFORMS::dst_frame /output_frame
---period 10 \
+--period 0.01 \
 --name /outlaser:o
 */
